Fix #864: Add string-ize operator '#'
Encloses the next word or number in double quotes. If the word is a macro, it is expanded before quoting.
A space is required between the hash and the word.
Removed constant expression operator from z80asm (#) - no longer required and collides with new string-ize operator.
Fixed the libraries still using '#' to start expressions with parenthesis, to differenciate from indirect addressing - replaced by a '+' sign.